# user-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "User profile and cohort management service",
        "title": "User Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/cohorts": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the list of cohorts available to the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cohorts"
                ],
                "summary": "List cohorts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.CohortResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new cohort owned by the authenticated teacher.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cohorts"
                ],
                "summary": "Create cohort",
                "parameters": [
                    {
                        "description": "Create cohort request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CohortCreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CohortResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cohorts/join": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Joins the authenticated user to a cohort identified by an invite token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cohorts"
                ],
                "summary": "Join cohort",
                "parameters": [
                    {
                        "description": "Join cohort request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CohortJoinRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cohorts/{id}/invite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates an invite token for a cohort owned by the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cohorts"
                ],
                "summary": "Generate invite token",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cohort ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/cohorts/{id}/members": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns cohort details and members if the authenticated user is the owner or a member.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cohorts"
                ],
                "summary": "Get cohort members",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Cohort ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CohortWithUsersResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/internal/cohorts/can-edit": {
            "post": {
                "security": [
                    {
                        "InternalToken": []
                    }
                ],
                "description": "Checks whether the supplied user owns the supplied cohort.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Check cohort ownership",
                "parameters": [
                    {
                        "description": "Ownership check request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CohortIsOwnedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/internal/cohorts/is-user-in": {
            "post": {
                "security": [
                    {
                        "InternalToken": []
                    }
                ],
                "description": "Checks whether the supplied user belongs to any of the supplied cohorts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "internal"
                ],
                "summary": "Check cohort membership",
                "parameters": [
                    {
                        "description": "Membership check request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CohortIsUserInRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CohortIsUserInResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the current user profile, creating the user record if it does not exist yet.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "Get current user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CohortCreateRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "Name of the cohort to create",
                    "type": "string",
                    "example": "Advanced Mathematics"
                }
            }
        },
        "dto.CohortIsOwnedRequest": {
            "type": "object",
            "required": [
                "cohort_id",
                "user_id"
            ],
            "properties": {
                "cohort_id": {
                    "description": "Cohort identifier to check ownership for",
                    "type": "string",
                    "example": "123"
                },
                "user_id": {
                    "description": "User identifier to check ownership of",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "dto.CohortIsUserInRequest": {
            "type": "object",
            "required": [
                "cohort_ids",
                "user_id"
            ],
            "properties": {
                "cohort_ids": {
                    "description": "List of cohort identifiers to check membership in",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        123,
                        124,
                        125
                    ]
                },
                "user_id": {
                    "description": "User identifier to check membership for",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "dto.CohortIsUserInResponse": {
            "type": "object",
            "properties": {
                "cohort_ids": {
                    "description": "Cohort IDs that the user is a member of",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "example": [
                        123,
                        124
                    ]
                }
            }
        },
        "dto.CohortJoinRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "description": "JWT token generated by cohort owner to invite users",
                    "type": "string",
                    "example": "eyJhbGc..."
                }
            }
        },
        "dto.CohortResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Cohort's unique identifier",
                    "type": "string",
                    "example": "123"
                },
                "name": {
                    "description": "Cohort's name",
                    "type": "string",
                    "example": "Advanced Mathematics"
                }
            }
        },
        "dto.CohortWithUsersResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "Cohort's unique identifier",
                    "type": "string",
                    "example": "123"
                },
                "name": {
                    "description": "Cohort's name",
                    "type": "string",
                    "example": "Advanced Mathematics"
                },
                "users": {
                    "description": "List of users who are members of this cohort",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.UserResponse"
                    }
                }
            }
        },
        "dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "errors": {
                    "description": "Error message or description",
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "display_name": {
                    "description": "User's display name",
                    "type": "string",
                    "example": "John Doe"
                },
                "id": {
                    "description": "User's unique identifier (UUID)",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "preferences": {
                    "description": "User's preferences stored as arbitrary JSON data",
                    "type": "object"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "InternalToken": {
            "type": "apiKey",
            "name": "Internal-Token",
            "in": "header"
        }
    }
}