# user-service/internal/dto/cohort.go
package dto

// CohortResponse represents a cohort's basic information.
//
// swagger:model
type CohortResponse struct {
	// Cohort's unique identifier
	ID string `json:"id" example:"123"`
	// Cohort's name
	Name string `json:"name" example:"Advanced Mathematics"`
}

// CohortWithUsersResponse represents a cohort with its members.
//
// swagger:model
type CohortWithUsersResponse struct {
	// Cohort's unique identifier
	ID string `json:"id" example:"123"`
	// Cohort's name
	Name string `json:"name" example:"Advanced Mathematics"`
	// List of users who are members of this cohort
	Users []UserResponse `json:"users"`
}

// CohortCreateRequest represents a request to create a new cohort.
//
// swagger:model
type CohortCreateRequest struct {
	// Name of the cohort to create
	Name string `json:"name" example:"Advanced Mathematics" binding:"required"`
}

// CohortJoinRequest represents a request to join a cohort using an invite token.
//
// swagger:model
type CohortJoinRequest struct {
	// JWT token generated by cohort owner to invite users
	Token string `json:"token" example:"eyJhbGc..." binding:"required"`
}

// CohortIsOwnedRequest represents a request to check cohort ownership.
// Used by internal services to verify if a user owns a specific cohort.
//
// swagger:model
type CohortIsOwnedRequest struct {
	// Cohort identifier to check ownership for
	CohortID string `json:"cohort_id" example:"123" binding:"required"`
	// User identifier to check ownership of
	UserID string `json:"user_id" example:"550e8400-e29b-41d4-a716-446655440000" binding:"required"`
}

// CohortIsUserInRequest represents a request to check cohort membership.
// Used by internal services to check if a user belongs to any of the specified cohorts.
//
// swagger:model
type CohortIsUserInRequest struct {
	// User identifier to check membership for
	UserID string `json:"user_id" example:"550e8400-e29b-41d4-a716-446655440000" binding:"required"`
	// List of cohort identifiers to check membership in
	CohortIDs []int64 `json:"cohort_ids" example:"123,124,125" binding:"required"`
}

// CohortIsUserInResponse represents the response for cohort membership check.
// Contains the list of cohort IDs that the user is a member of.
//
// swagger:model
type CohortIsUserInResponse struct {
	// Cohort IDs that the user is a member of
	CohortIDs []int64 `json:"cohort_ids" example:"123,124"`
}